requirements.txt
# Core dependencies - All FREE
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
# Note: sqlite3 is included with Python

# Data processing
pandas==2.1.3
requests==2.31.0

# Optional performance improvements
aiofiles==23.2.1
httpx==0.25.1
python-multipart==0.0.6
